# Rune extensions micro-optimizations