--- a/Benchmarks/Text/RuneExtensions/DecodeSurrogatePair.cs
+++ b/Benchmarks/Text/RuneExtensions/DecodeSurrogatePair.cs
@@ -12,51 +12,45 @@ namespace Terminal.Gui.Benchmarks.Text.RuneExtensions;
 public class DecodeSurrogatePair
 {
     /// <summary>
-    /// Benchmark for current ToString -> ToCharArray implementation.
-    /// </summary>
-    [Benchmark (Baseline = true)]
-    [ArgumentsSource (nameof (DataSource))]
-    public char [] Current (Rune rune)
-    {
-        Tui.RuneExtensions.DecodeSurrogatePair (rune, out char []? chars);
-        return chars;
-    }
-
-    /// <summary>
-    /// Benchmark for new implementation.
+    /// Benchmark for previous implementation.
     /// </summary>
     /// <param name="rune"></param>
     /// <returns></returns>
     [Benchmark]
     [ArgumentsSource (nameof (DataSource))]
-    public char []? New (Rune rune)
+    public char []? Previous (Rune rune)
     {
-        RuneEncodeToCharBuffer (rune, out char []? chars);
+        _ = RuneToStringToCharArray (rune, out char []? chars);
         return chars;
     }
 
     /// <summary>
+    /// Benchmark for current implementation.
+    /// 
     /// Utilizes Rune methods that take Span argument avoiding intermediate heap array allocation when combined with stack allocated intermediate buffer.
     /// When rune is not surrogate pair there will be no heap allocation.
     /// 
     /// Final surrogate pair array allocation cannot be avoided due to the current method signature design.
     /// Changing the method signature, or providing an alternative method, to take a destination Span would allow further optimizations by allowing caller to reuse buffer for consecutive calls.
     /// </summary>
-    private static bool RuneEncodeToCharBuffer (Rune rune, out char []? chars)
+    [Benchmark (Baseline = true)]
+    [ArgumentsSource (nameof (DataSource))]
+    public char []? Current (Rune rune)
     {
-        bool isSingleUtf16CodeUnit = rune.IsBmp;
-        if (isSingleUtf16CodeUnit)
-        {
-            chars = null;
-            return false;
-        }
+        _ = Tui.RuneExtensions.DecodeSurrogatePair (rune, out char []? chars);
+        return chars;
+    }
 
-        const int maxRuneUtf16Length = 2;
-        Span<char> charBuffer = stackalloc char[maxRuneUtf16Length];
-        int charsWritten = rune.EncodeToUtf16 (charBuffer);
-        if (charsWritten >= 2 && char.IsSurrogatePair (charBuffer [0], charBuffer [1]))
+    /// <summary>
+    /// Previous implementation with intermediate string allocation.
+    /// 
+    /// The IsSurrogatePair implementation at the time had hidden extra string allocation so there were intermediate heap allocations even if rune is not surrogate pair.
+    /// </summary>
+    private static bool RuneToStringToCharArray (Rune rune, out char []? chars)
+    {
+        if (rune.IsSurrogatePair ())
         {
-            chars = charBuffer [..charsWritten].ToArray ();
+            chars = rune.ToString ().ToCharArray ();
             return true;
         }