--- a/Benchmarks/Text/RuneExtensions/GetEncodingLength.cs
+++ b/Benchmarks/Text/RuneExtensions/GetEncodingLength.cs
@@ -16,7 +16,7 @@ public class GetEncodingLength
     /// </summary>
     [Benchmark]
     [ArgumentsSource (nameof (DataSource))]
-    public int Previous (Rune rune, Encoding encoding)
+    public int Previous (Rune rune, PrettyPrintedEncoding encoding)
     {
         return WithEncodingGetBytesArray (rune, encoding);
     }
@@ -26,7 +26,7 @@ public class GetEncodingLength
     /// </summary>
     [Benchmark (Baseline = true)]
     [ArgumentsSource (nameof (DataSource))]
-    public int Current (Rune rune, Encoding encoding)
+    public int Current (Rune rune, PrettyPrintedEncoding encoding)
     {
         return Tui.RuneExtensions.GetEncodingLength (rune, encoding);
     }
@@ -50,15 +50,25 @@ public class GetEncodingLength
 
     public static IEnumerable<object []> DataSource ()
     {
-        Encoding[] encodings = [ Encoding.UTF8, Encoding.Unicode, Encoding.UTF32 ];
+        PrettyPrintedEncoding[] encodings = [ new(Encoding.UTF8), new(Encoding.Unicode), new(Encoding.UTF32) ];
         Rune[] runes = [ new Rune ('a'), "𝔹".EnumerateRunes ().Single () ];
 
-        foreach (Encoding encoding in encodings)
+        foreach (var encoding in encodings)
         {
             foreach (Rune rune in runes)
             {
                 yield return [rune, encoding];
             }
         }
+    }
+
+    /// <summary>
+    /// <see cref="System.Text.Encoding"/> wrapper to display proper encoding name in benchmark results.
+    /// </summary>
+    public record PrettyPrintedEncoding (Encoding Encoding)
+    {
+        public static implicit operator Encoding (PrettyPrintedEncoding ppe) => ppe.Encoding;
+
+        public override string ToString () => Encoding.HeaderName;
     }
 }