## Fixes

Main motivation is to introduce basic benchmark project as framework for future performance fine-tuning.

- Improves #2725 at minuscule/theoretical level.
- Microbenchmarks and micro-optimizations for several RuneExtensions methods.
   - Theoretically the optimizations could improve performance of scenarios with heavy use of TextView but I doubt there is any practically difference.
- Marked RuneExtensions.MaxUnicodeCodePoint readonly.
- Regression test cases for RuneExtensions.GetEncodingLength.

## Proposed Changes/Todos

- [x] Enable null analysis in RuneExtensions
    - Potentially new warnings elsewhere.
    - DecodeSurrogatePair out parameter marked as potentially null.
- [x] Make RuneExtensions.MaxUnicodeCodePoint read-only
    - I guess technically breaking change.
- [x] Add minimal test cases for RuneExtensions.GetEncodingLength.
- [x] Add Terminal.Gui.Benchmarks project.
- [x] Add microbenchmarks for 5 RuneExtensions methods.
    - [x] DecodeSurrogate 
    - [x] Encode
    - [x] EncodeSurrogatePair
    - [x] GetEncodingLength
    - [x] IsSurrogatePair
- [x] Add micro-optimizations for 4 RuneExtension methods.
    - [x] DecodeSurrogate 
    - [x] Encode
        - **BREAKING CHANGE** due to count argument vs rune byte count check, the new implementation does not throw exception when the count argument is larger than the rune byte count. I'm not sure how critical this is in grand scheme.
    - [x] GetEncodingLength
    - [x] IsSurrogatePair

## Pull Request Checklist

- [ ] I've named my PR in the form of "Fixes #issue. Terse description."
   - *N/A, this PR does not "fix" any existing issue.*
- [x] My code follows the [style guidelines of Terminal.Gui](https://github.com/gui-cs/Terminal.Gui/blob/develop/.editorconfig) - if you use Visual Studio, hit `CTRL-K-D` to automatically reformat your files before committing.
- [x] My code follows the [Terminal.Gui library design guidelines](https://github.com/gui-cs/Terminal.Gui/blob/develop/CONTRIBUTING.md)
- [x] I ran `dotnet test` before commit
- [x] I have made corresponding changes to the API documentation (using `///` style comments)
- [ ] My changes generate no new warnings
  - *Enabling null analysis in RuneExtensions might have introduced warnings elsewhere...*
- [x] I have checked my code and corrected any poor grammar or misspellings
- [x] I conducted basic QA to assure all features are working

## Optimizations & Benchmark Results

Main motivation for optimizations for me was minimizing intermediate allocations. Any speed improvements were nice added bonus.

The benchmark arguments combinations/permutations are mainly limited to keep the measurement runs relatively short.

Results may vary greatly depending on computer specs. Benchmarks were run in Release build.

```
dotnet run --configuration Release --project Benchmarks/Benchmarks.csproj
```

Overall the micro-optimizations involve replacing methods using intermediate strings with similar methods that take (stack allocated) Span buffers.

Previous in the benchmark result tables means the original implementation before optimizations and current means the new optimized implementation. In case table only lists current then that means the original implementation was not touched.

BenchmarkDotNet environment summary:
```
BenchmarkDotNet v0.14.0, Windows 10 (10.0.19045.5487/22H2/2022Update)
AMD Ryzen 7 5800X3D, 1 CPU, 16 logical and 8 physical cores
.NET SDK 9.0.200
  [Host]     : .NET 8.0.13 (8.0.1325.6609), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.13 (8.0.1325.6609), X64 RyuJIT AVX2
```

### DecodeSurrogate

Replaced the string and char array allocations with stack allocated char buffer.

The difference is smaller than originally due to compounding optimizations from IsSurrogatePair that is used in the original/previous implementation.

| Method   | rune | Mean      | Error     | StdDev    | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|--------- |----- |----------:|----------:|----------:|------:|--------:|-------:|----------:|------------:|
| **Previous** | **a**    |  **2.038 ns** | **0.0398 ns** | **0.0311 ns** |  **1.49** |    **0.03** |      **-** |         **-** |          **NA** |
| Current  | a    |  1.368 ns | 0.0239 ns | 0.0200 ns |  1.00 |    0.02 |      - |         - |          NA |
|          |      |           |           |           |       |         |        |           |             |
| **Previous** | **𝔹**   | **10.228 ns** | **0.2504 ns** | **0.2342 ns** |  **1.28** |    **0.04** | **0.0013** |      **64 B** |        **2.00** |
| Current  | 𝔹   |  8.024 ns | 0.1899 ns | 0.1777 ns |  1.00 |    0.03 | 0.0006 |      32 B |        1.00 |

### Encode

Replaced string and byte array allocations with stack allocated byte buffer.

| Method   | rune | destination | start | count | Mean      | Error     | StdDev    | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|--------- |----- |------------ |------ |------ |----------:|----------:|----------:|------:|--------:|-------:|----------:|------------:|
| **Previous** | **a**    | **Byte[16]**    | **0**     | **-1**    | **14.935 ns** | **0.0478 ns** | **0.0373 ns** |  **4.07** |    **0.02** | **0.0011** |      **56 B** |          **NA** |
| Current  | a    | Byte[16]    | 0     | -1    |  3.671 ns | 0.0144 ns | 0.0128 ns |  1.00 |    0.00 |      - |         - |          NA |
|          |      |             |       |       |           |           |           |       |         |        |           |             |
| **Previous** | **a**    | **Byte[16]**    | **8**     | **-1**    | **15.389 ns** | **0.3513 ns** | **0.3286 ns** |  **3.88** |    **0.13** | **0.0011** |      **56 B** |          **NA** |
| Current  | a    | Byte[16]    | 8     | -1    |  3.967 ns | 0.1178 ns | 0.1102 ns |  1.00 |    0.04 |      - |         - |          NA |
|          |      |             |       |       |           |           |           |       |         |        |           |             |
| **Previous** | **𝔞**   | **Byte[16]**    | **0**     | **-1**    | **19.457 ns** | **0.0560 ns** | **0.0468 ns** |  **3.05** |    **0.01** | **0.0013** |      **64 B** |          **NA** |
| Current  | 𝔞   | Byte[16]    | 0     | -1    |  6.375 ns | 0.0108 ns | 0.0091 ns |  1.00 |    0.00 |      - |         - |          NA |
|          |      |             |       |       |           |           |           |       |         |        |           |             |
| **Previous** | **𝔞**   | **Byte[16]**    | **8**     | **-1**    | **19.706 ns** | **0.2950 ns** | **0.2759 ns** |  **3.08** |    **0.04** | **0.0013** |      **64 B** |          **NA** |
| Current  | 𝔞   | Byte[16]    | 8     | -1    |  6.390 ns | 0.0254 ns | 0.0212 ns |  1.00 |    0.00 |      - |         - |          NA |

### EncodeSurrogatePair

After IsSurrogatePair optimizations, it seemed good enough because there were no heap allocations anymore so I didn't touch the implementation. It mostly already does same stuff as Rune internally.

Also this is not really used anywhere outside tests.

| Method  | highSurrogate | lowSurrogate | Mean     | Error     | StdDev    | Ratio | RatioSD | Allocated | Alloc Ratio |
|-------- |-------------- |------------- |---------:|----------:|----------:|------:|--------:|----------:|------------:|
| **Current** | **\ud83c**        | **\udf39**       | **1.358 ns** | **0.0531 ns** | **0.0029 ns** |  **1.00** |    **0.00** |         **-** |          **NA** |
|         |               |              |          |           |           |       |         |           |             |
| **Current** | **\ud83c**        | **\udf55**       | **1.362 ns** | **0.1688 ns** | **0.0093 ns** |  **1.00** |    **0.01** |         **-** |          **NA** |
|         |               |              |          |           |           |       |         |           |             |
| **Current** | **\ud83e**        | **\udde0**       | **1.361 ns** | **0.5898 ns** | **0.0323 ns** |  **1.00** |    **0.03** |         **-** |          **NA** |

### GetEncodingLength

Replaced string, char array, and byte array allocations with stack allocated char and byte buffers. Char buffer is first used to extract chars from rune. After that the byte buffer with given encoding is used for determining the final encoding length.

| Method   | rune | encoding | Mean      | Error     | StdDev    | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|--------- |----- |--------- |----------:|----------:|----------:|------:|--------:|-------:|----------:|------------:|
| **Previous** | **a**    | **utf-16**   | **19.688 ns** | **0.2498 ns** | **0.2086 ns** |  **2.02** |    **0.04** | **0.0017** |      **88 B** |          **NA** |
| Current  | a    | utf-16   |  9.742 ns | 0.1891 ns | 0.1769 ns |  1.00 |    0.02 |      - |         - |          NA |
|          |      |          |           |           |           |       |         |        |           |             |
| **Previous** | **a**    | **utf-32**   | **44.649 ns** | **0.6810 ns** | **0.6371 ns** |  **2.01** |    **0.03** | **0.0062** |     **312 B** |        **2.79** |
| Current  | a    | utf-32   | 22.162 ns | 0.1689 ns | 0.1410 ns |  1.00 |    0.01 | 0.0022 |     112 B |        1.00 |
|          |      |          |           |           |           |       |         |        |           |             |
| **Previous** | **a**    | **utf-8**    | **24.061 ns** | **0.2509 ns** | **0.1959 ns** |  **3.17** |    **0.03** | **0.0017** |      **88 B** |          **NA** |
| Current  | a    | utf-8    |  7.594 ns | 0.0108 ns | 0.0084 ns |  1.00 |    0.00 |      - |         - |          NA |
|          |      |          |           |           |           |       |         |        |           |             |
| **Previous** | **𝔹**   | **utf-16**   | **24.130 ns** | **0.4747 ns** | **0.7105 ns** |  **2.11** |    **0.06** | **0.0019** |      **96 B** |          **NA** |
| Current  | 𝔹   | utf-16   | 11.426 ns | 0.0167 ns | 0.0156 ns |  1.00 |    0.00 |      - |         - |          NA |
|          |      |          |           |           |           |       |         |        |           |             |
| **Previous** | **𝔹**   | **utf-32**   | **47.434 ns** | **0.5403 ns** | **0.4790 ns** |  **2.03** |    **0.02** | **0.0063** |     **320 B** |        **2.86** |
| Current  | 𝔹   | utf-32   | 23.346 ns | 0.1244 ns | 0.0971 ns |  1.00 |    0.01 | 0.0022 |     112 B |        1.00 |
|          |      |          |           |           |           |       |         |        |           |             |
| **Previous** | **𝔹**   | **utf-8**    | **26.614 ns** | **0.1977 ns** | **0.1753 ns** |  **2.57** |    **0.02** | **0.0019** |      **96 B** |          **NA** |
| Current  | 𝔹   | utf-8    | 10.360 ns | 0.0283 ns | 0.0221 ns |  1.00 |    0.00 |      - |         - |          NA |

### IsSurrogatePair

Replaced string allocation with stack allocated char buffer. Also added early return by eliminating single code point runes with the built-in rune property (`Rune.IsBmp`) for checking Basic Multilingual Plane (BMP) range.

| Method   | rune | Mean      | Error     | StdDev    | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|--------- |----- |----------:|----------:|----------:|------:|--------:|-------:|----------:|------------:|
| **Previous** | **a**    | **3.4912 ns** | **0.0339 ns** | **0.0283 ns** |  **5.66** |    **0.36** | **0.0005** |      **24 B** |          **NA** |
| Current  | a    | 0.6190 ns | 0.0376 ns | 0.0418 ns |  1.00 |    0.09 |      - |         - |          NA |
|          |      |           |           |           |       |         |        |           |             |
| **Previous** | **𝔹**   | **2.8274 ns** | **0.0552 ns** | **0.0489 ns** |  **2.77** |    **0.12** | **0.0006** |      **32 B** |          **NA** |
| Current  | 𝔹   | 1.0212 ns | 0.0444 ns | 0.0455 ns |  1.00 |    0.06 |      - |         - |          NA |