@tig 2025-02-25
I'll try to get to this today. 

In the meantime, can you run the UICatalog benchmarks with and without this and share the results?