@tig 2025-02-25
```ps1
.\UICatalog\bin\Release\net8.0\UICatalog.exe -b
```

With your  PR:

![image](https://github.com/user-attachments/assets/c4afd15d-eb38-4593-a6b0-15b2a82898e7)


Without:

![image](https://github.com/user-attachments/assets/0ea2d706-7c26-45e3-be1a-630f31bf4fd8)