@TheTonttu 2025-02-25
Oh, I didn't even know there is that sort of benchmark. 🤦‍♂️
Good to know for the future.