@TheTonttu 2025-02-25
Also as far as I know the rune extensions are only utilized in TextView so would be interesting to look more closely at a scenario that use those specifically.