@tig 2025-02-25
> Oh, I didn't even know there is that sort of benchmark. 🤦‍♂️ Good to know for the future.

I added it fairly recently. It's a bit clunky so I'd love feedback. If you look at Scenario.cs in UICatalog you'll see that there's the ability to "script" scenarios to make them sing & dance as part of this, which I think can be super useful for benchmarking at the scenario level.